Refactor and relocate WildfireAntivirusProfile factories
Moved WildfireAntivirusProfile-related factories to a dedicated module `tests/test_factories/security/wildfire_antivirus_profile.py` for improved modularity and organization. Updated imports across relevant test files to reflect the new structure.
